move tool definitions to new plugin
