+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/create_visualization/guess_chart_type.ts
