+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/create_visualization/prompts.ts
