+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/execute_esql.ts
