+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/generate_esql.ts
