+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/get_document_by_id.ts
