+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/get_index_mapping.ts
