+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/index_explorer.ts
