+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/list_indices.ts
