+++ x-pack/platform/plugins/shared/onechat/server/plugin.ts
@@ -24,7 +24,6 @@ import { registerOnechatHandlerContext } from './request_handler_context';
 import { createOnechatUsageCounter } from './telemetry/usage_counters';
 import { TrackingService } from './telemetry/tracking_service';
 import { registerTelemetryCollector } from './telemetry/telemetry_collector';
-import { registerBuiltinTools } from './services/tools';
 
 export class OnechatPlugin
   implements
@@ -75,12 +74,6 @@ export class OnechatPlugin
     registerUISettings({ uiSettings: coreSetup.uiSettings });
 
     registerOnechatHandlerContext({ coreSetup });
-
-    registerBuiltinTools({
-      registry: serviceSetups.tools,
-      coreSetup,
-      setupDeps,
-    });
 
     const router = coreSetup.http.createRouter<OnechatHandlerContext>();
     registerRoutes({
