+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/index.ts
@@ -6,5 +6,4 @@
  */
 
 export { createBuiltinToolRegistry, type BuiltinToolRegistry } from './builtin_registry';
-export { registerBuiltinTools } from './register_tools';
 export { createBuiltinProviderFn } from './builtin_provider';
